# swagger/spring-server/src/main/resources/api-spec.yaml
swagger: "2.0"
info:
  description: "This is the specification for the gamification project"
  version: "1.0.0"
  title: "Swagger Gamification"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "vincent.guidoux1@heig-vd.ch"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost"
basePath: "/api"
tags:
  - name: "badges"
    description: "Everything about the badges"
  - name: "pointScales"
    description: "Everything about the pointScales"
  - name: "rules"
    description: "Everything about the rules"
  - name: "users"
    description: "Everything about the users"
  - name: "events"
    description: "Everything about the event"
schemes:
  - "http"
parameters:
  token:
    in: header
    name: authorization
    type: string
    required: true
  badgeId:
    in: path
    name: badgeId
    type: integer
    format: int64
    required: true
  userId:
    in: path
    name: userId
    type: integer
    format: int64
    required: true
  pointScaleId:
    in: path
    name: pointScaleId
    type: integer
    format: int64
    required: true
  ruleId:
    in: path
    name: ruleId
    type: integer
    format: int64
    required: true
paths:
  /badges:
    post:
      tags:
        - "badges"
      summary: "Creates a badge"
      description: "Creates a badge"
      operationId: "createBadge"
      produces:
        - "application/json"
      consumes:
        - "application/json"
      parameters:
        - $ref: '#/parameters/token'
        - in: "body"
          name: "body"
          description: "badge to create"
          schema:
            $ref: "#/definitions/badgeWrite"
      responses:
        200:
          description: "successful operation - here's your id"
          schema:
            type: string
        204:
          description: "Already exist !"
    get:
      tags:
        - "badges"
      summary: "Get all badges for one application"
      description: "Get all badges for one application"
      operationId: "getBadges"
      parameters:
        - $ref: '#/parameters/token'
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: array
            items:
              $ref: "#/definitions/badgeRead"
  /badges/{badgeId}:
    parameters:
      - $ref: '#/parameters/badgeId'
    get:
      tags:
        - "badges"
      summary: "Get badge by ID"
      description: "Get one badge of the application using its id"
      operationId: "getBadge"
      parameters:
        - $ref: '#/parameters/token'
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/badgeRead"
        401:
          description: "Unauthorized"
        404:
          description: "Not found"
    put:
      tags:
        - "badges"
      summary: "Update an existing badge"
      description: "Update an existing badge"
      operationId: "updateBadge"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - $ref: '#/parameters/token'
        - in: "body"
          name: "body"
          description: "badge that needs to be update in the store"
          schema:
            $ref: "#/definitions/badgeWrite"
      responses:
        204:
          description: "successful operation"
        401:
          description: "Unauthorized"
        404:
          description: "Object not found"
    delete:
      tags:
        - "badges"
      summary: "Deletes a badge"
      description: "Deletes a badge"
      operationId: "deleteBadge"
      produces:
        - "application/json"
      parameters:
        - $ref: '#/parameters/token'
      responses:
        204:
          description: "successfull operation"
        401:
          description: "Unauthorized"
        404:
          description: "object not found"

  /pointScales:
    post:
      tags:
        - "pointScales"
      summary: "Creates a point scale"
      description: "Creates a point scale"
      operationId: "createPointScale"
      produces:
        - "application/json"
      parameters:
        - $ref: '#/parameters/token'
        - in: "body"
          name: "body"
          description: "pointScale to create"
          schema:
            $ref: "#/definitions/pointScaleWrite"
      responses:
        200:
          description: "successful operation - here's your id"
          schema:
            type: string
        204:
          description: "Already exist !"
    get:
      tags:
        - "pointScales"
      summary: "Get all point scales for one application"
      description: "Get all point scales for one application"
      operationId: "getPointScales"
      parameters:
        - $ref: '#/parameters/token'
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: array
            items:
              $ref: "#/definitions/pointScaleRead"
  /pointScales/{pointScaleId}:
    get:
      tags:
        - "pointScales"
      summary: "Get a specific point scale for one application"
      description: "Get a specific point scale for one application"
      operationId: "getPointScale"
      parameters:
        - $ref: '#/parameters/token'
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/pointScaleRead"
        401:
          description: "Unauthorized"
        404:
          description: "Not found"
    parameters:
      - $ref: '#/parameters/pointScaleId'
    put:
      tags:
        - "pointScales"
      summary: "Update an existing point scale"
      description: "Update an existing point scale"
      operationId: "updatePointScale"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - $ref: '#/parameters/token'
        - in: "body"
          name: "body"
          description: "point scale that needs to be update in the store"
          schema:
            $ref: "#/definitions/pointScaleWrite"
      responses:
        204:
          description: "successful operation"
        401:
          description: "Unauthorized"
        404:
          description: "Object not found"
    delete:
      tags:
        - "pointScales"
      summary: "Deletes a point scale"
      description: "Deletes a point scale"
      operationId: "deletePointScale"
      produces:
        - "application/json"
      parameters:
        - $ref: '#/parameters/token'
      responses:
        204:
          description: "successfull operation"
        401:
          description: "Unauthorized"
        404:
          description: "object not found"
  /rules:
    post:
      tags:
        - "rules"
      summary: "Creates a rule"
      description: "Creates a rule"
      operationId: "createRule"
      produces:
        - "application/json"
      parameters:
        - $ref: '#/parameters/token'
        - in: "body"
          name: "body"
          description: "rule to create"
          schema:
            $ref: "#/definitions/ruleWrite"
      responses:
        200:
          description: "successful operation - here's your id"
          schema:
            type: string
        204:
          description: "Already exist !"
    get:
      tags:
        - "rules"
      summary: "Get all rules for one application"
      description: "Get all rules for one application"
      operationId: "getRules"
      parameters:
        - $ref: '#/parameters/token'
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: array
            items:
              $ref: "#/definitions/ruleRead"
  /rules/{ruleId}:
    parameters:
      - $ref: '#/parameters/ruleId'
    get:
      tags:
        - "rules"
      summary: "Get a specific rule for one application"
      description: "Get a specific rule for one application"
      operationId: "getRule"
      parameters:
        - $ref: '#/parameters/token'
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ruleRead"
        401:
          description: "Unauthorized"
        404:
          description: "Not found"
    put:
      tags:
        - "rules"
      summary: "Update an existing rule"
      description: "Update an existing rule"
      operationId: "updateRule"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - $ref: '#/parameters/token'
        - in: "body"
          name: "body"
          description: "rule that needs to be update in the store"
          schema:
            $ref: "#/definitions/ruleWrite"
      responses:
        204:
          description: "successful operation"
        401:
          description: "Unauthorized"
        404:
          description: "Object not found"
    delete:
      tags:
        - "rules"
      summary: "Deletes a rule"
      description: "Deletes a rule"
      operationId: "deleteRule"
      produces:
        - "application/json"
      parameters:
        - $ref: '#/parameters/token'
      responses:
        204:
          description: "successfull operation"
        401:
          description: "Unauthorized"
        404:
          description: "object not found"
  /events:
    post:
      tags:
        - "events"
      summary: "Creates an event"
      description: "Creates an event"
      operationId: "createEvent"
      produces:
        - "application/json"
      parameters:
        - $ref: '#/parameters/token'
        - in: "body"
          name: "body"
          description: "event to create"
          schema:
            $ref: "#/definitions/event"
      responses:
        200:
          description: "successful operation"
  /users/{userId}/badges:
    parameters:
      - $ref: '#/parameters/userId'
    get:
      tags:
        - "users"
      summary: "Get all badges from a specific user"
      description: "Get all badges from a specific user"
      operationId: "getBadgesFromUser"
      parameters:
        - $ref: '#/parameters/token'
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: array
            items:
              $ref: "#/definitions/badgeRead"
  /users/{userId}/pointScales:
    parameters:
      - $ref: '#/parameters/userId'
    get:
      tags:
        - "users"
      summary: "Get all pointScales from a specific user"
      description: "Get all pointScales from a specific user"
      operationId: "getPointScalesFromUser"
      parameters:
        - $ref: '#/parameters/token'
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: array
            items:
              $ref: '#/definitions/pointScaleRead'

definitions:
  pointScaleWrite:
    type: object
    required:
      - name
      - description
    properties:
      name:
        type: string
      description:
        type: string
  pointScaleRead:
    type: object
    required:
      - name
      - description
      - id
    properties:
      name:
        type: string
      description:
        type: string
      id:
        type: integer
        format: int64
  badgeWrite:
    type: object
    required:
      - name
      - image
      - visible
    properties:
      name:
        type: string
      image:
        type: string
      visible:
        type: boolean
  badgeRead:
    type: object
    required:
      - name
      - image
      - visible
      - id
    properties:
      name:
        type: string
      image:
        type: string
      visible:
        type: boolean
      id:
        type: integer
        format: int64
  ruleWrite:
    type: object
    required:
      - ifEventType
    properties:
      ifEventType:
        type: string
      ifPropertyName:
        type: string
      ifPropertyCondition:
        type: string
      thenBadgeId:
        type: integer
        format: int64
      thenPointScaleId:
        type: integer
        format: int64
      thenAwardPoint:
        type: integer
  ruleRead:
    type: object
    required:
      - id
      - ifEventType
    properties:
      id:
        type: integer
        format: int64
      ifEventType:
        type: string
      ifPropertyName:
        type: string
      ifPropertyCondition:
        type: string
      thenBadgeId:
        type: integer
        format: int64
      thenPointScaleId:
        type: integer
        format: int64
      thenAwardPoint:
        type: integer
  event:
    type: object
    required:
      - userId
      - type
    properties:
      userId:
        type: integer
        format: int64
      type:
        type: string
      propertyName:
        type: string
      propertyDefinition:
        type: string